+++ testapp/src/main/java/com/rakuten/tech/mobile/testapp/ui/userdata/ContactAdapterPresenter.kt
@@ -1,0 +1,8 @@
+package com.rakuten.tech.mobile.testapp.ui.userdata
+
+interface ContactAdapterPresenter {
+    fun addContact(position: Int, contact: String)
+    fun addContactList(contacts: ArrayList<String>)
+    fun removeContactAt(position: Int)
+    fun provideContactEntries(): ArrayList<String>
+}